+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- `create` and `publish` commands now print created/published file paths using the Laravel `make:*` style output (`components->info` + `twoColumnDetail`).
+- `git config user.name` and `git symbolic-ref --short HEAD` lookups now use `Illuminate\Support\Facades\Process` instead of `exec` for testability.
+
 ## [0.4.0] - 0.4.0
+
 - PHP 8.3 and 8.4 support added.
 - PHPUnit 11 and 12 support added.
 - Laravel Zero upgraded to 12.0.
@@ -11,14 +16,18 @@ All notable changes to this project will be documented in this file.
 - Fix namespacing issues.
 
 ## [0.3.2] - 0.3.2
+
 - PHP 8.2 support added.
 
 ## [0.3.1] - 0.3.1
+
 - Fix namespacing issues
 
 ## [0.3.0] - 0.3.0
+
 - PHP 8.1 support added.
 - Upgrade to Laravel Zero 9.0.0
 
 ## [0.2.0] - 0.2.0
+
 - PHP 8.0 support added.